Created services for Order and Customer along with Tests.

+++ TechTest/AnyCompany.Data/Order.cs
@@ -3,8 +3,8 @@
     public class Order
     {
         public int OrderId { get; set; }
-        public double Amount { get; set; }
-        public double VAT { get; set; }
+        public decimal Amount { get; set; }
+        public decimal VAT { get; set; }
         public int CustomerId { get; set; }
     }
 }
